+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 ## [Unreleased]
 ### Added
 - option to hide legend
+- option for logarithmic y-axis
 
 ## [0.11.0]
 ### Added